ethicode/requirements.txt
reflex==0.4.0
sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel